+++ pkg/interpreter/src/defaults.js
@@ -1,0 +1,17 @@
+import Regex from './library/regex'
+
+const defaults = {
+  assoc () { return {} },
+  boolean () { return false },
+  integer () { return 0 },
+  list () { return [] },
+  long () { return 0 },
+  real () { return 0 },
+  regex () { return new Regex() },
+  string () { return '' }
+}
+
+export default function getDefaultValue (type) {
+  const getter = defaults[type]
+  return getter && getter()
+}