+++ pkg/interpreter/src/interpreter.js
@@ -1,26 +1,47 @@
 import Scope from './scope'
+import getDefaultValue from './defaults'
 
 const push = Array.prototype.push
 
+class InterpreterError extends Error {}
+class NotImplementedError extends InterpreterError {}
+class RuntimeError extends InterpreterError {
+  constructor (message, node) {
+    super(message)
+    this.node = node
+  }
+}
+
 function ignore () {}
 
-const interpreter = {}
+let warnings
 
-export default interpreter
+export function setOptions (options = {}) {
+  ({ warnings } = options)
+}
 
-interpreter.Program = (node, scope, walk) => walk(node.body, node, scope)
+export const visitors = {}
+
+// ---------- Program
+
+visitors.Program = (node, scope, walk) => walk(node.body, node, scope)
 
 // ---------- Package
 
-interpreter.PackageDeclaration = () => {
-  throw new Error('package is not implemented')
+visitors.PackageDeclaration = () => {
+  throw new NotImplementedError('package is not implemented')
 }
 
-interpreter.ObjectName = ({ raw }, scope) => scope.get(raw)
+visitors.ObjectName = (node, scope) => {
+  let { raw } = node
+  if (!raw) raw = prepareObjectName(node)
+  return scope.get(raw)
+}
 
 // ---------- Script
 
-interpreter.ScriptSource = (node, scope, walk) => {
+visitors.ScriptSource = (node, scope, walk) => {
+  if (!node.statements) prepareScript(node)
   for (const routine of node.routines) walk(routine, node, scope)
   for (const variable of node.variables) walk(variable, node, scope)
   for (const statement of node.statements) walk(statement, node, scope)
@@ -28,19 +49,25 @@ interpreter.ScriptSource = (node, scope, walk) => {
 
 // ---------- Dump
 
-interpreter.DumpSource = () => {
-  throw new Error('dump is not implemented')
+visitors.DumpSource = () => {
+  throw new NotImplementedError('dump is not implemented')
 }
 
 // ---------- Scopes
 
-interpreter.ScriptDeclaration = interpreter.FunctionDeclaration = (node, scope) => {
+visitors.ScriptDeclaration = (node, scope) => {
+  if (!node.statements) prepareScript(node)
+  scope.setOwn(node.id.value, node)
+}
+
+visitors.FunctionDeclaration = (node, scope) => {
+  if (!node.statements) prepareFunction(node)
   scope.setOwn(node.id.value, node)
 }
 
 // ---------- Statements
 
-interpreter.IfStatement = (node, scope, walk) => {
+visitors.IfStatement = (node, scope, walk) => {
   if (walk(node.test, node, scope)) {
     for (const statement of node.consequent) walk(statement, node, scope)
   } else {
@@ -58,7 +85,7 @@ function tryElseIfClauses (otherClauses, scope, walk) {
   }
 }
 
-interpreter.SwitchStatement = (node, scope, walk) => {
+visitors.SwitchStatement = (node, scope, walk) => {
   const discriminant = walk(node.discriminant, node, scope)
   node.cases.some(switchCase => trySwitchCase(discriminant, switchCase, scope, walk))
 }
@@ -70,7 +97,7 @@ function trySwitchCase (discriminant, node, scope, walk) {
   }
 }
 
-interpreter.WhileStatement = (node, scope, walk) => {
+visitors.WhileStatement = (node, scope, walk) => {
   while (walk(node.test, node, scope)) {
     scope.break = scope.continue = false
     for (const statement of node.body) {
@@ -81,7 +108,7 @@ interpreter.WhileStatement = (node, scope, walk) => {
   }
 }
 
-interpreter.RepeatStatement = (node, scope, walk) => {
+visitors.RepeatStatement = (node, scope, walk) => {
   do {
     scope.break = scope.continue = false
     for (const statement of node.body) {
@@ -92,7 +119,7 @@ interpreter.RepeatStatement = (node, scope, walk) => {
   } while (walk(node.test, node, scope))
 }
 
-interpreter.ForStatement = (node, scope, walk) => {
+visitors.ForStatement = (node, scope, walk) => {
   for (node.init && walk(node.init, node, scope);
     !node.test || walk(node.test, node, scope);
     node.update && walk(node.update, node, scope)) {
@@ -105,7 +132,7 @@ interpreter.ForStatement = (node, scope, walk) => {
   }
 }
 
-interpreter.ForEachStatement = (node, scope, walk) => {
+visitors.ForEachStatement = (node, scope, walk) => {
   const name = node.left.value
   for (const item of walk(node.right, node, scope)) {
     scope.set(name, item)
@@ -118,7 +145,7 @@ interpreter.ForEachStatement = (node, scope, walk) => {
   }
 }
 
-interpreter.StructuredForStatement = (node, scope, walk) => {
+visitors.StructuredForStatement = (node, scope, walk) => {
   const name = node.variable.value
   let index = walk(node.start, node, scope)
   const end = walk(node.end, node, scope)
@@ -134,58 +161,60 @@ interpreter.StructuredForStatement = (node, scope, walk) => {
   }
 }
 
-interpreter.BreakStatement = (node, scope) => { scope.break = true }
+visitors.BreakStatement = (node, scope) => { scope.break = true }
 
-interpreter.ContinueStatement = (node, scope) => { scope.continue = true }
+visitors.ContinueStatement = (node, scope) => { scope.continue = true }
 
-interpreter.BreakIfStatement = (node, scope, walk) => {
+visitors.BreakIfStatement = (node, scope, walk) => {
   scope.break = walk(node.test, node, scope)
 }
 
-interpreter.ContinueIfStatement = (node, scope, walk) => {
+visitors.ContinueIfStatement = (node, scope, walk) => {
   scope.continue = walk(node.test, node, scope)
 }
 
-interpreter.GotoStatement = (node, scope, walk) => {
-  throw new Error('package is not implemented')
+visitors.GotoStatement = (node, scope, walk) => {
+  throw new NotImplementedError('goto is not implemented')
 }
 
-interpreter.LabelStatement = interpreter.EmptyStatement = ignore
+visitors.LabelStatement = visitors.EmptyStatement = ignore
 
-interpreter.ReturnStatement = (node, scope, walk) => {
+visitors.ReturnStatement = (node, scope, walk) => {
   if (node.argument) scope.result = walk(node.argument, node, scope)
 }
 
-interpreter.VariableDeclaration = (node, scope, walk) => {
+visitors.VariableDeclaration = (node, scope, walk) => {
   for (const declaration of node.declarations) {
     const { init } = declaration
-    scope.setOwn(declaration.id.value, init && walk(init, declaration, scope))
+    scope.setOwn(declaration.id.value,
+      init ? walk(init, declaration, scope) : getDefaultValue(node.variableType))
   }
 }
 
 // ---------- Expressions
 
-interpreter.ConditionalExpression = (node, scope, walk) => {
+visitors.ConditionalExpression = (node, scope, walk) => {
   return walk(node.test, node, scope)
     ? walk(node.consequent, node, scope)
     : walk(node.alternate, node, scope)
 }
 
-interpreter.BinaryExpression = (node, scope, walk) => {
-  let { left, operator, right } = node
-  left = isAssignment(operator) ? getName(left) : walk(left, node, scope)
-  return binaryOperation[operator](left, walk(right, node, scope))
+visitors.BinaryExpression = (node, scope, walk) => {
+  const { left, operator, right } = node
+  let target
+  if (isAssignment(operator)) {
+    // TODO: Implement other left operands than an identifier.
+    target = getName(left)
+    if (!target) throw new RuntimeError('assignment target is not a variable', left)
+  } else {
+    target = walk(left, node, scope)
+  }
+  return binaryOperation[operator](target, walk(right, node, scope), scope, left)
 }
 
 function isAssignment (symbol) {
-  switch (symbol.length) {
-    case 1:
-      return symbol === '='
-    case 2:
-      return symbol === '+=' || symbol === '-=' || symbol === '*=' ||
-        symbol === '&=' || symbol === '|=' || symbol === '^='
-  }
-  return false
+  return symbol === '=' ||  symbol === '+=' || symbol === '-=' ||
+    symbol === '*=' || symbol === '&=' || symbol === '|=' || symbol === '^='
 }
 
 function getName (node) {
@@ -194,11 +223,18 @@ function getName (node) {
     ? node.value
     : type === 'ObjectName' || type === 'LegacyAlias'
       ? node.raw
-      : 'dynamic'
+      : undefined
+}
+
+function setVariable (name, value, scope, target) {
+  if (!scope.set(name, value)) {
+    if (warnings) throw new RuntimeError('variable not declared', target)
+    scope.setOwn(name, value)
+  }
 }
 
 const binaryOperation = {
-  '=': (left, right, scope) => scope.set(left, right),
+  '=': (left, right, scope, target) => setVariable(left, right, scope, target),
   '+': (left, right) => left + right,
   '-': (left, right) => left - right,
   '*': (left, right) => left * right,
@@ -215,14 +251,14 @@ const binaryOperation = {
   '<=': (left, right) => left <= right,
   '>=': (left, right) => left >= right,
   '&&': (left, right) => left && right,
+  '||': (left, right) => left || right,
   in: (left, right) => left in right,
-  '+=': (left, right, scope) => scope.set(left, scope.get(left) + right),
-  '-=': (left, right, scope) => scope.set(left, scope.get(left) - right),
-  '*=': (left, right, scope) => scope.set(left, scope.get(left) * right),
-  '&=': (left, right, scope) => scope.set(left, scope.get(left) & right),
-  '|=': (left, right, scope) => scope.set(left, scope.get(left) | right),
-  '^=': (left, right, scope) => scope.set(left, scope.get(left) ^ right),
-  '^^': (left, right) => !!(left ^ right),
+  '+=': (left, right, scope, target) => setVariable(left, scope.get(left) + right, scope, target),
+  '-=': (left, right, scope, target) => setVariable(left, scope.get(left) - right, scope, target),
+  '*=': (left, right, scope, target) => setVariable(left, scope.get(left) * right, scope, target),
+  '&=': (left, right, scope, target) => setVariable(left, scope.get(left) & right, scope, target),
+  '|=': (left, right, scope, target) => setVariable(left, scope.get(left) | right, scope, target),
+  '^=': (left, right, scope, target) => setVariable(left, scope.get(left) ^ right, scope, target),
   '<<': (left, right) => left << right,
   '>>': (left, right) => left >> right,
   or: (left, right) => left || right,
@@ -236,7 +272,7 @@ const binaryOperation = {
   xor: (left, right) => !!(left ^ right)
 }
 
-interpreter.UnaryExpression = (node, scope, walk) =>
+visitors.UnaryExpression = (node, scope, walk) =>
   unaryOperation[node.operator](walk(node.argument, node, scope))
 
 const unaryOperation = {
@@ -247,29 +283,60 @@ const unaryOperation = {
   not: value => !value
 }
 
-interpreter.MemberExpression = (node, scope, walk) =>
-  walk(node.object, node, scope)[walk(node.property, node, scope)]
+visitors.MemberExpression = (node, scope, walk) => {
+  const object = walk(node.object, node, scope)
+  if (!object) throw new RuntimeError('object is undefined', node.object)
+  if (!(typeof object === 'object' && object && !Array.isArray(object))) {
+    throw new RuntimeError(`dereferenced source (${typeof array}) is not an object or assoc`, node.object)
+  }
+  let { property } = node
+  const { type } = property
+  property = type === 'Identifier' || type === 'LegacyAlias'
+    ? property.value
+    : walk(node.property, node, scope)
+  return object[property]
+}
 
-interpreter.SliceExpression = (node, scope, walk) => {
+visitors.SliceExpression = (node, scope, walk) => {
   const array = walk(node.object, node, scope)
+  if (!array) throw new RuntimeError('slice source is undefined', node.object)
+  if (!(typeof array === 'string' || Array.isArray(array))) {
+    throw new RuntimeError(`slice source (${typeof array}) is not a list or string`, node.object)
+  }
   const start = (node.start && walk(node.start, node, scope)) || 0
   const end = (node.end && walk(node.end, node, scope)) || array.length
   return array.slice(start, end)
 }
 
-interpreter.IndexExpression = (node, scope, walk) =>
-  walk(node.object, node, scope)[walk(node.index, node, scope)]
+visitors.IndexExpression = (node, scope, walk) => {
+  const array = walk(node.object, node, scope)
+  if (!array) throw new RuntimeError('index source is undefined', node.object)
+  if (!(typeof array === 'string' || Array.isArray(array))) {
+    throw new RuntimeError(`index source (${typeof array}) is not a list or string`, node.object)
+  }
+  return array[walk(node.index, node, scope)]
+}
 
-interpreter.CallExpression = (node, scope, walk) => {
+visitors.CallExpression = (node, scope, walk) => {
   const routine = walk(node.callee, node, scope)
-  if (!routine) throw new Error(`routine "${getName(node.callee)}" is undefined`)
+  if (!routine) {
+    const name = getName(node.callee)
+    if (name) throw new RuntimeError(`function "${name}" is undefined`, node.callee)
+    throw new RuntimeError('undefined is not a function', node.callee)
+  }
   const values = node.arguments.map(arg => walk(arg, node, scope))
   if (typeof routine === 'function') return routine(...values)
-  if (!routine.type) throw new Error(`"${getName(node.callee)}" is not a routine`)
+  if (!routine.type) {
+    const name = getName(node.callee)
+    if (name) throw new RuntimeError(`"${name}" is not a function`, node.callee)
+    throw new RuntimeError(`${typeof routine} is not a function`, node.callee)
+  }
   const args = routine.params.reduce((args, param, index) => {
     const value = values[index]
-    args[param.id.value] = value === undefined && param.init
-      ? walk(param.init, node, scope)
+    args[param.id.value] = value === undefined
+      ? param.init
+          ? walk(param.init, node, scope)
+          : getDefaultValue(param.parameterType)
       : value
     return args
   }, {})
@@ -281,18 +348,17 @@ interpreter.CallExpression = (node, scope, walk) => {
   return scope.result
 }
 
-interpreter.ThisExpression = (node, scope) => scope
+visitors.ThisExpression = (node, scope) => scope
 
-interpreter.SuperExpression = (node, { upper }) => upper
+visitors.SuperExpression = (node, { upper }) => upper
 
-interpreter.AssocExpression = interpreter.ObjectExpression = (node, scope, walk) => {
-  return node.properties.reduce((object, { key, value }) => {
+visitors.AssocExpression = visitors.ObjectExpression = (node, scope, walk) =>
+  node.properties.reduce((object, { key, value }) => {
     object[walk(key, node, scope)] = walk(value, node, scope)
     return object
   }, {})
-}
 
-interpreter.ListExpression = (node, scope, walk) =>
+visitors.ListExpression = (node, scope, walk) =>
   node.elements.reduce((result, element) => {
     const item = walk(element, node, scope)
     if (element.type === 'AtExpression') push.apply(result, item)
@@ -300,7 +366,7 @@ interpreter.ListExpression = (node, scope, walk) =>
     return result
   }, [])
 
-interpreter.ListComprehension = (node, scope, walk) => {
+visitors.ListComprehension = (node, scope, walk) => {
   const name = node.left.value
   const values = walk(node.right, node, scope)
   return values.reduce((result, value) => {
@@ -312,15 +378,70 @@ interpreter.ListComprehension = (node, scope, walk) => {
   }, [])
 }
 
-interpreter.AtExpression = interpreter.ParenthesisExpression =
+visitors.AtExpression = visitors.ParenthesisExpression =
   (node, scope, walk) => walk(node.expression, node, scope)
 
-interpreter.XlateExpression = ({ raw }, scope) => scope.get(raw)
+visitors.XlateExpression = (node, scope) => {
+  let { raw } = node
+  if (!raw) raw = prepareXlateExpression(node)
+  scope.get(raw)
+}
 
 // ---------- Identifiers and Literals
 
-interpreter.Identifier = ({ value }, scope) => scope.get(value)
+visitors.Identifier = ({ value }, scope) => scope.get(value)
 
-interpreter.LegacyAlias = ({ raw }, scope) => scope.get(raw)
+visitors.LegacyAlias = (node, scope) => {
+  let { raw } = node
+  if (!raw) raw = prepareLegacyAlias(node)
+  scope.get(raw)
+}
 
-interpreter.Literal = ({ value }) => value
+visitors.Literal = ({ value }) => value
+
+// ---------- Preparation
+
+function prepareScript (node) {
+  const routines = node.routines = []
+  const variables = node.variables = []
+  const statements = node.statements = []
+  for (const part of node.body) {
+    const { type } = part
+    const target = type === 'VariableDeclaration'
+      ? variables
+      : type === 'FunctionDeclaration' || type === 'ScriptDeclaration'
+        ? routines
+        : statements
+    target.push(part)
+  }
+}
+
+function prepareFunction (node) {
+  const variables = node.variables = []
+  const statements = node.statements = []
+  for (const part of node.body) {
+    const { type } = part
+    const target = type === 'VariableDeclaration' ? variables : statements
+    target.push(part)
+  }
+}
+
+function prepareObjectName (node) {
+  const raw = node.name
+    .map(({ value }) => typeof value === 'number' ? `&${value.toString(16)}` : value)
+    .join('::')
+  node.raw = raw
+  return raw
+}
+
+function prepareLegacyAlias (node) {
+  const raw = `&${node.value.toString(16)}`
+  node.raw = raw
+  return raw
+}
+
+function prepareXlateExpression (node) {
+  const raw = `${node.ospace.value}.${node.string.value}`
+  node.raw = raw
+  return raw
+}